Deductible absence in payslips details
